ref(ddm): Rename tray content to widget details
