+++ static/app/views/ddm/widgetDetails.tsx
